Bring in the ExceptionHelper
This generalizes exception handling, also improved handling of umbrella exceptions.

Refactored the CommandEventBus a bit to reduce duplication, and fix a bug where if the firing of a Command fails, the deferred commands will no longer be fired, causing the eventbus and application to essentially stop working.

Added a GuardedEventBus which logs errors that happen during event execution. This is useful because Vue's global error handler (or other error handlers) typically take precedent when something throws an exception during execution. This behaviour (which is intended) can be preserved by using the CommandEventBus, or can be handled by using the GuardedEventBus instead.