Fix VuelidateUtil
With new version of elemental it seems the proxy method in VuelidateUtil broke.
It uses JSObject.assign to assing the validations object to an JsPropertyMap.
In the old implementation this resulted in an entry at key "0".
But in the new elemental it assigns the value to obj.
And than get("0") will return undefined.

The code was changed to pass the value directly without using JsPropertyMap and JsObject assign (not sure why that was used in the first place).